refactor(portfolio): 프리랜서 경력, 외부 프로젝트 Controller, Service 분리

+++ src/main/java/com/sanaiclub/portfolio/dao/FreelancerCareerMapper.java
@@ -14,6 +14,6 @@ public interface FreelancerCareerMapper {
 
     int updateCareer(FreelancerCareerVO vo);
 
-    int deleteCareer(@Param("careerId") Long careerId,
+    int deleteCareer(@Param("careerId") Integer careerId,
                      @Param("freelancerId") Integer freelancerId);
 }
